feat: rename and removed body
